requirements.txt
appdirs==1.4.4
attrs==21.1.0
black==21.5b0
certifi==2020.12.5
click==7.1.2
dj-database-url==0.5.0
Django==3.2.10
django-on-heroku==1.1.2
dropbox==10.10.0
gunicorn==20.1.0
jsonschema==3.2.0
mypy-extensions==0.4.3
numpy==1.21
pathspec==0.8.1
pyrsistent==0.17.3
pytz==2021.1
regex==2021.4.4
scipy==1.6.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.2.0
pylint==2.12.2
python-decouple==3.5
